Commit 18:  DB Debuggin 99% - new deleteByID fun: using the "id" (Long), no more the text "q" ! (well, now when I have got the an ID in the "FavoriteQuote" class, I should use it consequeontly) - maybe I can use both funs in order to eliminate Doublettes !?!??!!?

+++ app/src/main/java/com/example/spiritualguide01/home/quotes/datamodel/FavoriteQuote.kt
@@ -8,5 +8,5 @@ data class FavoriteQuote(
     val q: String,
     val a: String,
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0
+    val id: Long = 0L
 )
